adding AI stuff to seperate AI package
